Remove shared changes and only implement endpoint through VGS
