+++ forage-android/src/main/java/com/joinforage/forage/android/collect/BTPinCollector.kt
@@ -20,21 +20,13 @@ import com.verygoodsecurity.vgscollect.core.HTTPMethod
 import org.json.JSONException
 import java.util.*
 
+// IMPORTANT: Any changes to this object must be reflected in the consumer-rules.pro file
 internal data class ProxyRequestObject(val pin: TextElement, val card_number_token: String)
-
-internal data class POSProxyRequestObject(
-    val pin: TextElement,
-    val card_number_token: String,
-    val amount: String,
-    val reason: String,
-    val pos_terminal: Map<String, Any>,
-    val metadata: Map<String, Any>?
-)
 
 internal class BTPinCollector(
     private val pinForageEditText: ForagePINEditText,
     private val merchantAccount: String
-) : PinCollector() {
+) : PinCollector {
     private val logger = Log.getInstance()
     private val vaultType = VaultType.BT_VAULT_TYPE
 
@@ -61,7 +53,7 @@ internal class BTPinCollector(
         val measurement = setupMeasurement(balancePath(paymentMethodRef), UserAction.BALANCE)
 
         val proxyRequest: ProxyRequest = ProxyRequest().apply {
-            headers = buildBTHeaders(encryptionKey, merchantAccount, traceId = logger.getTraceIdValue())
+            headers = buildHeaders(encryptionKey, merchantAccount, traceId = logger.getTraceIdValue())
             body = ProxyRequestObject(
                 pin = pinForageEditText.getTextElement(),
                 card_number_token = cardToken
@@ -151,7 +143,7 @@ internal class BTPinCollector(
         val measurement = setupMeasurement(capturePaymentPath(paymentRef), UserAction.CAPTURE)
 
         val proxyRequest: ProxyRequest = ProxyRequest().apply {
-            headers = buildBTHeaders(encryptionKey, merchantAccount, paymentRef, traceId = logger.getTraceIdValue())
+            headers = buildHeaders(encryptionKey, merchantAccount, paymentRef, traceId = logger.getTraceIdValue())
             body = ProxyRequestObject(
                 pin = pinForageEditText.getTextElement(),
                 card_number_token = cardToken
@@ -241,7 +233,7 @@ internal class BTPinCollector(
         val bt = buildBt()
 
         val proxyRequest: ProxyRequest = ProxyRequest().apply {
-            headers = buildBTHeaders(encryptionKey, merchantAccount, paymentRef, traceId = logger.getTraceIdValue())
+            headers = buildHeaders(encryptionKey, merchantAccount, paymentRef, traceId = logger.getTraceIdValue())
             body = ProxyRequestObject(
                 pin = pinForageEditText.getTextElement(),
                 card_number_token = cardToken
@@ -314,105 +306,6 @@ internal class BTPinCollector(
         )
     }
 
-    override suspend fun submitPosRefund(
-        paymentRef: String,
-        cardToken: String,
-        encryptionKey: String,
-        terminalId: String,
-        amount: String,
-        reason: String,
-        metadata: Map<String, Any>?
-    ): ForageApiResponse<String> {
-        // TODO: Maybe change the logged attributes
-        val logAttributes = mapOf(
-            "merchant_ref" to merchantAccount,
-            "payment_ref" to paymentRef
-        )
-        // If the PIN isn't valid (less than 4 numbers) then return a response here.
-        if (!pinForageEditText.getElementState().isComplete) {
-            return returnIncompletePinError(logAttributes, logger)
-        }
-
-        val bt = buildBt()
-
-        val proxyRequest: ProxyRequest = ProxyRequest().apply {
-            headers = buildBTHeaders(encryptionKey, merchantAccount, paymentRef, traceId = logger.getTraceIdValue())
-            body = POSProxyRequestObject(
-                pin = pinForageEditText.getTextElement(),
-                card_number_token = cardToken,
-                amount = amount,
-                metadata = metadata,
-                reason = reason,
-                pos_terminal = mapOf("provider_terminal_id" to terminalId)
-            )
-            path = posRefundPath(paymentRef)
-        }
-        val measurement = setupMeasurement(posRefundPath(paymentRef), UserAction.REFUND)
-
-        logger.i(
-            "[BT] Sending POS refund to BasisTheory",
-            attributes = logAttributes
-        )
-
-        measurement.start()
-        val response = runCatching {
-            bt.proxy.post(proxyRequest)
-        }
-        measurement.end()
-
-        // MUST reset the PIN value after submitting
-        pinForageEditText.getTextElement().setText("")
-
-        if (response.isSuccess) {
-            val forageResponse = response.getOrNull()
-            try {
-                val forageApiError = ForageApiError.ForageApiErrorMapper.from(forageResponse.toString())
-                val error = forageApiError.errors[0]
-                logger.e(
-                    "[BT] Received an error while submitting POS refund request to BasisTheory: $error.message",
-                    attributes = logAttributes
-                )
-
-                // Error code hardcoded as 400 because of lack of information
-                val httpStatusCode = 400
-                measurement.setHttpStatusCode(httpStatusCode).setForageErrorCode(error.code).logResult()
-                return ForageApiResponse.Failure(
-                    listOf(
-                        ForageError(
-                            httpStatusCode,
-                            error.code,
-                            error.message
-                        )
-                    )
-                )
-            } catch (_: JSONException) { }
-            logger.i(
-                "[BT] Received successful response from BasisTheory",
-                attributes = logAttributes
-            )
-
-            // Status Code hardcoded because of lack of knowledge
-            val httpStatusCode = 200
-            measurement.setHttpStatusCode(httpStatusCode).logResult()
-
-            return ForageApiResponse.Success(forageResponse.toString())
-        }
-        val btErrorResponse = response.exceptionOrNull()
-        logger.e(
-            "[BT] Received BasisTheory API exception while calling submitPosRefund: $btErrorResponse",
-            attributes = logAttributes
-        )
-
-        val unknownBtStatusCode = 500
-        measurement.setHttpStatusCode(unknownBtStatusCode).logResult()
-
-        return ForageApiResponse.Failure(
-            listOf(
-                ForageError(unknownBtStatusCode, "unknown_server_error", "Unknown Server Error")
-            )
-        )
-    }
-
     override fun parseEncryptionKey(encryptionKeys: EncryptionKeys): String {
         return encryptionKeys.btAlias
     }
@@ -442,24 +335,6 @@ internal class BTPinCollector(
             .setMethod(HTTPMethod.POST.toString())
     }
 
-    private fun buildBTHeaders(
-        encryptionKey: String,
-        merchantAccount: String,
-        idempotencyKey: String = UUID.randomUUID().toString(),
-        traceId: String = ""
-    ): Map<String, String> {
-        val headers = buildHeaders(
-            encryptionKey,
-            merchantAccount,
-            idempotencyKey,
-            traceId
-        )
-        headers[ForageConstants.Headers.BT_PROXY_KEY] = PROXY_ID
-        headers[ForageConstants.Headers.CONTENT_TYPE] = "application/json"
-
-        return headers
-    }
-
     companion object {
         // this code assumes that .setForageConfig() has been called
         // on a Forage***EditText before PROXY_ID or API_KEY get
@@ -473,6 +348,22 @@ internal class BTPinCollector(
                 .build()
         }
 
+        private fun buildHeaders(
+            encryptionKey: String,
+            merchantAccount: String,
+            idempotencyKey: String = UUID.randomUUID().toString(),
+            traceId: String = ""
+        ): Map<String, String> {
+            val headers = HashMap<String, String>()
+            headers[ForageConstants.Headers.X_KEY] = encryptionKey
+            headers[ForageConstants.Headers.MERCHANT_ACCOUNT] = merchantAccount
+            headers[ForageConstants.Headers.IDEMPOTENCY_KEY] = idempotencyKey
+            headers[ForageConstants.Headers.BT_PROXY_KEY] = PROXY_ID
+            headers[ForageConstants.Headers.CONTENT_TYPE] = "application/json"
+            headers[ForageConstants.Headers.TRACE_ID] = traceId
+            return headers
+        }
+
         internal fun returnIncompletePinError(logAttributes: Map<String, String>, logger: Log): ForageApiResponse.Failure {
             logger.w(
                 "[BT] User attempted to submit an invalid PIN",
@@ -482,5 +373,14 @@ internal class BTPinCollector(
                 ForageConstants.ErrorResponseObjects.INCOMPLETE_PIN_ERROR
             )
         }
+
+        fun balancePath(paymentMethodRef: String) =
+            "/api/payment_methods/$paymentMethodRef/balance/"
+
+        fun capturePaymentPath(paymentRef: String) =
+            "/api/payments/$paymentRef/capture/"
+
+        fun deferPaymentCapturePath(paymentRef: String) =
+            "/api/payments/$paymentRef/collect_pin/"
     }
 }
